requirements.txt
transformers
torch
sentencepiece
python-dotenv
requests
streamlit
Pillow
